[Frontend] Renamed cardItem and mainLayout files with camelCase
